Medic/requirements.txt
flask>=3.0.0
Cerberus>=1.3.5
psycopg2-binary>=2.9.9
pytz>=2024.1
slack-sdk>=3.26.0
requests>=2.31.0
flask_swagger_ui>=4.11.1
prometheus-client>=0.19.0
